Update project image
I always miss/forget the project image. Is there a way to have it be part of the "new package" setup process? Such as just a box with a label on the "Add Package" page of OpenUPM? Simply seeing it during the setup process would certainly help in remembering that it exists.