Update Git for Unity readme, description and image
Git for Unity packages are automatically generated and published into release branches, and they were missing the top level readme until now, so when it was registered in openupm, the only readme available was for a subpackage, which was wrong. This fixes that, and also adds an image link and a cleaner description.